+++ src/io/ft_putuint_base.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putuint_base.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iwordes <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: iwordes <iwordes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/14 19:40:59 by iwordes           #+#    #+#             */
-/*   Updated: 2016/11/14 19:41:01 by iwordes          ###   ########.fr       */
+/*   Updated: 2016/12/14 19:10:55 by iwordes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ft_io.h"
+#include <libft.h>
 
-void	ft_putuint_base(unsigned long long integer, int base)
+void	ft_putuint_base(uintmax_t integer, int base)
 {
 	ft_putuint_base_fd(integer, base, 1);
 }
